Fix serialization tests (#3)
The multi-line JSON fragments in the code have CRLF.  String
comparison fails on Mac and Linux because the left and right side
of Assert.AreEqual on have different line-endings. Converting the
code to use LF makes them pass on Linux and fail on Windows.

I added a TestUtils.Normalize function to convert CRLF to LF.
This is now used to wrap expected value.